# pkg/fs: optimize GetObject, add benchmark