This PR optimizes `GetObject` by delaying the excess system calls unless they are actually necessary. They're used to give the correct error message back to the user, but the common case is that the correct object is retrieved, and not the bucket is not found. This optimizes for the common case. This should not penalize the uncommon case, as no system calls are added, it just restructures them so as to avoid them when unnecessary.

It also adds a benchmark so you can see the improvement (on my OSX laptop, drops ~500 ns/op):
```
# With patch
BenchmarkGetObject-4    	   30000	     48197 ns/op
BenchmarkGetObject-4    	   30000	     47472 ns/op
BenchmarkGetObject-4    	   30000	     44748 ns/op
# Without patch
BenchmarkGetObject-4    	   30000	     52736 ns/op
BenchmarkGetObject-4    	   30000	     52875 ns/op
BenchmarkGetObject-4    	   30000	     53476 ns/op
```