@brendanashworth 2016-03-13
Looks like there's a difference in `Seek` for windows and unix — pushing a fix for that now.