@brendanashworth 2016-03-13
@harshavardhana I've changed the `err` to `e` and have added a `GOOS` check, PTAL!